requirements.txt
# Core Web Framework
streamlit==1.40.0

# Data Processing
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
xlrd==2.0.1

# Visualization
plotly==5.24.1
matplotlib==3.9.2
seaborn==0.13.2

# Machine Learning
scikit-learn==1.5.2
xgboost==2.1.3
lightgbm==4.5.0
imbalanced-learn==0.12.4

# Time Series Forecasting
prophet==1.1.6

# AI Integration
google-generativeai==0.8.3

# Google Cloud Storage
google-cloud-storage==2.18.2

# PDF Generation
reportlab==4.2.5
fpdf2==2.8.1

# Utilities
python-dotenv==1.0.1
Pillow==11.0.0
scipy==1.14.1
joblib==1.4.2

# Custom Additions
kaleido==0.2.1
patool==4.0.4
rarfile==4.2
openai==1.55.3
